moved metadata endpoints and test to separate files and general tidy up for common code
