+++ api/dataset_test.go
@@ -3,13 +3,10 @@ package api
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"testing"
 	"time"
 
@@ -40,11 +37,7 @@ var (
 	errBadRequest = errors.New("bad request")
 	errNotFound   = errors.New("not found")
 
-	datasetPayload           = `{"contacts":[{"email":"testing@hotmail.com","name":"John Cox","telephone":"01623 456789"}],"description":"census","links":{"access_rights":{"href":"http://ons.gov.uk/accessrights"}},"title":"CensusEthnicity","theme":"population","periodicity":"yearly","state":"completed","next_release":"2016-04-04","publisher":{"name":"The office of national statistics","type":"government department","url":"https://www.ons.gov.uk/"}}`
-	editionPayload           = `{"edition":"2017","state":"created"}`
-	versionPayload           = `{"instance_id":"a1b2c3","edition":"2017","license":"ONS","release_date":"2017-04-04"}`
-	versionAssociatedPayload = `{"instance_id":"a1b2c3","edition":"2017","license":"ONS","release_date":"2017-04-04","state":"associated","collection_id":"12345"}`
-	versionPublishedPayload  = `{"instance_id":"a1b2c3","edition":"2017","license":"ONS","release_date":"2017-04-04","state":"published","collection_id":"12345"}`
+	datasetPayload = `{"contacts":[{"email":"testing@hotmail.com","name":"John Cox","telephone":"01623 456789"}],"description":"census","links":{"access_rights":{"href":"http://ons.gov.uk/accessrights"}},"title":"CensusEthnicity","theme":"population","periodicity":"yearly","state":"completed","next_release":"2016-04-04","publisher":{"name":"The office of national statistics","type":"government department","url":"https://www.ons.gov.uk/"}}`
 
 	urlBuilder                    = url.NewBuilder("localhost:20000")
 	genericMockedObservationStore = &mocks.ObservationStoreMock{}
@@ -856,431 +849,6 @@ func TestDeleteDatasetReturnsError(t *testing.T) {
 		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 0)
 		So(len(mockedDataStore.DeleteDatasetCalls()), ShouldEqual, 0)
 	})
-}
-
-func TestGetMetadataReturnsOk(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully return metadata resource for a request without an authentication header", t, func() {
-
-		datasetDoc := createDatasetDoc()
-		versionDoc := createVersionDoc()
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetID, edition, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, edition, version, state string) (*models.Version, error) {
-				return versionDoc, nil
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-
-		bytes, err := ioutil.ReadAll(w.Body)
-		if err != nil {
-			os.Exit(1)
-		}
-
-		var metaData models.Metadata
-
-		err = json.Unmarshal(bytes, &metaData)
-		if err != nil {
-			os.Exit(1)
-		}
-
-		So(metaData.Keywords, ShouldBeNil)
-		So(metaData.ReleaseFrequency, ShouldEqual, "yearly")
-
-		temporal := models.TemporalFrequency{
-			EndDate:   "2017-05-09",
-			Frequency: "Monthly",
-			StartDate: "2014-05-09",
-		}
-		So(metaData.Temporal, ShouldResemble, &[]models.TemporalFrequency{temporal})
-		So(metaData.UnitOfMeasure, ShouldEqual, "Pounds Sterling")
-	})
-
-	// Subtle difference between the test above and below, keywords is Not nil
-	// in response for test below while it is nil in test above
-	Convey("Successfully return metadata resource for a request with an authentication header", t, func() {
-
-		datasetDoc := createDatasetDoc()
-		versionDoc := createVersionDoc()
-
-		r, err := createRequestWithAuth("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetID, edition, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, edition, version, state string) (*models.Version, error) {
-				return versionDoc, nil
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-
-		bytes, err := ioutil.ReadAll(w.Body)
-		if err != nil {
-			os.Exit(1)
-		}
-
-		var metaData models.Metadata
-
-		err = json.Unmarshal(bytes, &metaData)
-		if err != nil {
-			os.Exit(1)
-		}
-
-		So(metaData.Keywords, ShouldResemble, []string{"pensioners"})
-		So(metaData.ReleaseFrequency, ShouldResemble, "yearly")
-
-		temporal := models.TemporalFrequency{
-			EndDate:   "2017-05-09",
-			Frequency: "Monthly",
-			StartDate: "2014-05-09",
-		}
-		So(metaData.Temporal, ShouldResemble, &[]models.TemporalFrequency{temporal})
-		So(metaData.UnitOfMeasure, ShouldEqual, "Pounds Sterling")
-	})
-}
-
-func TestGetMetadataReturnsError(t *testing.T) {
-	t.Parallel()
-	Convey("When the api cannot connect to datastore return an internal server error", t, func() {
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return nil, errInternal
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, "internal error\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset document cannot be found for version return status not found", t, func() {
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return nil, errs.ErrDatasetNotFound
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldEqual, "Dataset not found\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the dataset document has no current sub document return status not found", t, func() {
-
-		datasetDoc := createDatasetDoc()
-		datasetDoc.Current = nil
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetId, edition, state string) error {
-				return nil
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldEqual, "Dataset not found\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the edition document cannot be found for version return status not found", t, func() {
-
-		datasetDoc := createDatasetDoc()
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetId, edition, state string) error {
-				return errs.ErrEditionNotFound
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldEqual, "Edition not found\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 0)
-	})
-
-	Convey("When the version document cannot be found return status not found", t, func() {
-
-		datasetDoc := createDatasetDoc()
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetId, edition, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, edition, version, state string) (*models.Version, error) {
-				return nil, errs.ErrVersionNotFound
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-		So(w.Body.String(), ShouldEqual, "Version not found\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-
-	Convey("When the version document state is invalid return an internal server error", t, func() {
-
-		datasetDoc := createDatasetDoc()
-
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/metadata", nil)
-		w := httptest.NewRecorder()
-
-		mockedDataStore := &storetest.StorerMock{
-			GetDatasetFunc: func(datasetID string) (*models.DatasetUpdate, error) {
-				return datasetDoc, nil
-			},
-			CheckEditionExistsFunc: func(datasetId, edition, state string) error {
-				return nil
-			},
-			GetVersionFunc: func(datasetID, edition, version, state string) (*models.Version, error) {
-				return &models.Version{State: "gobbly-gook"}, nil
-			},
-		}
-
-		api := GetAPIWithMockedDatastore(mockedDataStore, &mocks.DownloadsGeneratorMock{}, getMockAuditor(), genericMockedObservationStore)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-		So(w.Body.String(), ShouldEqual, "Incorrect resource state\n")
-
-		So(len(mockedDataStore.GetDatasetCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.CheckEditionExistsCalls()), ShouldEqual, 1)
-		So(len(mockedDataStore.GetVersionCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestCreateNewVersionDoc(t *testing.T) {
-	t.Parallel()
-	Convey("Check the version has the new collection id when request contains a collection_id", t, func() {
-		currentVersion := &models.Version{}
-		version := &models.Version{
-			CollectionID: "4321",
-		}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.CollectionID, ShouldNotBeNil)
-		So(version.CollectionID, ShouldEqual, "4321")
-	})
-
-	Convey("Check the version collection id does not get replaced by the current collection id when request contains a collection_id", t, func() {
-		currentVersion := &models.Version{
-			CollectionID: "1234",
-		}
-		version := &models.Version{
-			CollectionID: "4321",
-		}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.CollectionID, ShouldNotBeNil)
-		So(version.CollectionID, ShouldEqual, "4321")
-	})
-
-	Convey("Check the version has the old collection id when request is missing a collection_id", t, func() {
-		currentVersion := &models.Version{
-			CollectionID: "1234",
-		}
-		version := &models.Version{}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.CollectionID, ShouldNotBeNil)
-		So(version.CollectionID, ShouldEqual, "1234")
-	})
-
-	Convey("check the version collection id is not set when both request body and current version document are missing a collection id", t, func() {
-		currentVersion := &models.Version{}
-		version := &models.Version{}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.CollectionID, ShouldNotBeNil)
-		So(version.CollectionID, ShouldEqual, "")
-	})
-
-	Convey("Check the version has the new spatial link when request contains a links.spatial.href", t, func() {
-		currentVersion := &models.Version{}
-		version := &models.Version{
-			Links: &models.VersionLinks{
-				Spatial: &models.LinkObject{
-					HRef: "http://ons.gov.uk/geographylist",
-				},
-			},
-		}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.Links, ShouldNotBeNil)
-		So(version.Links.Spatial, ShouldNotBeNil)
-		So(version.Links.Spatial.HRef, ShouldEqual, "http://ons.gov.uk/geographylist")
-	})
-
-	Convey("Check the version links.spatial.href does not get replaced by the current version value", t, func() {
-		currentVersion := &models.Version{
-			Links: &models.VersionLinks{
-				Spatial: &models.LinkObject{
-					HRef: "http://ons.gov.uk/oldgeographylist",
-				},
-			},
-		}
-		version := &models.Version{
-			Links: &models.VersionLinks{
-				Spatial: &models.LinkObject{
-					HRef: "http://ons.gov.uk/geographylist",
-				},
-			},
-		}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.Links, ShouldNotBeNil)
-		So(version.Links.Spatial, ShouldNotBeNil)
-		So(version.Links.Spatial.HRef, ShouldEqual, "http://ons.gov.uk/geographylist")
-	})
-
-	Convey("Check the links.spatial.href has the old value when request does not contain a links.spatial.href", t, func() {
-		currentVersion := &models.Version{
-			Links: &models.VersionLinks{
-				Spatial: &models.LinkObject{
-					HRef: "http://ons.gov.uk/oldgeographylist",
-				},
-			},
-		}
-		version := &models.Version{}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.Links, ShouldNotBeNil)
-		So(version.Links.Spatial, ShouldNotBeNil)
-		So(version.Links.Spatial.HRef, ShouldEqual, "http://ons.gov.uk/oldgeographylist")
-	})
-
-	Convey("check the version links.spatial.href is not set when both request body and current version document do not contain a links.spatial.href", t, func() {
-		currentVersion := &models.Version{
-			Links: &models.VersionLinks{
-				Dataset: &models.LinkObject{
-					HRef: "http://ons.gov.uk/datasets/123",
-				},
-			},
-		}
-		version := &models.Version{}
-
-		populateNewVersionDoc(currentVersion, version)
-		So(version.Links, ShouldNotBeNil)
-		So(version.Links.Spatial, ShouldBeNil)
-	})
-}
-
-// createDatasetDoc returns a datasetUpdate doc containing minimal fields but
-// there is a clear difference between the current and next sub documents
-func createDatasetDoc() *models.DatasetUpdate {
-	generalDataset := &models.Dataset{
-		CollectionID:     "4321",
-		ReleaseFrequency: "yearly",
-		State:            models.PublishedState,
-		UnitOfMeasure:    "Pounds Sterling",
-	}
-
-	nextDataset := *generalDataset
-	nextDataset.CollectionID = "3434"
-	nextDataset.Keywords = []string{"pensioners"}
-	nextDataset.State = models.AssociatedState
-
-	datasetDoc := &models.DatasetUpdate{
-		ID:      "123",
-		Current: generalDataset,
-		Next:    &nextDataset,
-	}
-
-	return datasetDoc
-}
-
-func createVersionDoc() *models.Version {
-	temporal := models.TemporalFrequency{
-		EndDate:   "2017-05-09",
-		Frequency: "Monthly",
-		StartDate: "2014-05-09",
-	}
-	versionDoc := &models.Version{
-		State:        models.EditionConfirmedState,
-		CollectionID: "3434",
-		Temporal:     &[]models.TemporalFrequency{temporal},
-	}
-
-	return versionDoc
 }
 
 func getMockAuditor() *audit.AuditorServiceMock {
