+++ api/editions_test.go
@@ -15,6 +15,10 @@ import (
 	. "github.com/smartystreets/goconvey/convey"
 )
 
+const (
+	editionPayload = `{"edition":"2017","state":"created"}`
+)
+
 func TestGetEditionsReturnsOK(t *testing.T) {
 
 	t.Parallel()
